Fixed a minor bug

+++ byom.py
@@ -13,7 +13,7 @@ smoothing_value = sys.argv[3]
 
 training_filepath = "data/training-tweets.txt"
 testing_filepath = "data/test-tweets-given.txt"
-model = BigramLanguageModel(int(vocabulary), Decimal(penalty_weight), Decimal(smoothing_value))
+model = BigramLanguageModel(int(vocabulary), float(penalty_weight), Decimal(smoothing_value))
 print("\n--------Training Bi-gram Language Model--------")
 model.learnFromFile(training_filepath)
 print("\n--------Predicting From Test File--------")
